refactor: lift name comparison out of functions.cpp into a testable leaf module
