+++ src/deluge/util/functions.h
@@ -25,6 +25,7 @@
 #include "util/d_string.h"
 #include "util/fixedpoint.h"
 #include "util/lookuptables/lookuptables.h"
+#include "util/name_compare.h" // IWYU pragma: export - strcmpspecial and friends used to live here
 #include "util/waves.h"
 #include <bit>
 #include <cstdint>
@@ -327,15 +328,11 @@ inline q31_t sampleTriangleDistribution() {
 
 void seedRandom();
 
-extern bool shouldInterpretNoteNames;
-extern bool octaveStartsFromA;
-
 int32_t random(int32_t upperLimit);
 bool shouldDoPanning(int32_t panAmount, int32_t* amplitudeL, int32_t* amplitudeR);
 
 uint32_t getOscInitialPhaseForZero(OscType waveType);
 int32_t fastPythag(int32_t x, int32_t y);
-int32_t strcmpspecial(char const* first, char const* second);
 int32_t doLanczos(int32_t* data, int32_t pos, uint32_t posWithinPos, int32_t memoryNumElements);
 int32_t doLanczosCircular(int32_t* data, int32_t pos, uint32_t posWithinPos, int32_t memoryNumElements);
